Updated DB, EF model, wired API

+++ Vinder.WebAPI/Models/Vehicle.cs
@@ -8,7 +8,7 @@ namespace Vinder.WebAPI.Models
     public class Vehicle
     {
         public string Vin { get; set; }
-        public decimal Distance { get; set; }
+        public double Distance { get; set; }
         public string MasterImage { get; set; }
         public string[] Images { get; set; }
         public Dictionary<string, string> AttributesDictionary { get; set; }
